Added comments for Javadocs into src/main/java/com/comp2042/logic/bricks
Added comments into all files in the package to test Javadocs. Found out that all the Brick-shape-classes do not appear on the webpage due to having default access modifier of private. But when tested by adding a public modifier to the IBrick class, the class appears on the webpage of Javadocs.

Try changing all access modifiers to public/protected in the future and test again.